VISUS-12: Workaround fix for PROP_PRODUCT_VERSION
Yes it is truly a workaround as "unspecified" will also be patched because this is badly implemented by me :/
But for times sake this will work fine as every normal project will have a version set and only on edge cases (e.g. unit tests) this will break.
Updated documentation regarding this manifest attribute!