VISUS-13: Patching PROP_PRODUCT_VERSION improved
PROP_PRODUCT_VERSION will not be patched when PROP_PRODUCT_RC is correct RC, when is should not be patched (property missing or explicitely false) and when no version is set (defaults in Gradle to "unspecified").
Also fixed issue with PROP_BUILD_TIME now only showing hours and minutes and adding PROP_BUILD_DATE - documentation updated based on the changes.
Changed the system property for the ticked id from "JIRA_TICKET" to "TICKET_ID" - documentation updated based on the changes.
Fixed some minor issues with JAR extension method "patchManifest" and the corresponding test cases, now the logic works even when the task destination directory was changed.
Enabled some test cases disabled prior due to issues with the patching of PROP_PRODUCT_VERSION which was now resolved in this commit.